REFACTOR: 수정 요청 사항 수정 완료
