--- public/data/PRODUCTSORTS.js
@@ -1,7 +1,0 @@
-export const PRODUCTSORTS = [
-  { id: 1, name: '전체', link: '/product-page' },
-  { id: 2, name: '돼지', link: '/product-page' },
-  { id: 3, name: '소', link: '/product-page' },
-  { id: 4, name: '닭', link: '/product-page' },
-  { id: 5, name: '오리', link: '/product-page' },
-];
