+++ src/components/ProductListTable/ProductListTable.scss
@@ -1,4 +1,5 @@
 @import url('https://fonts.googleapis.com/css2?family=Noto+Sans+KR&display=swap');
+
 .ProductListLayout {
   display: grid;
   grid-template-columns: 400px 400px 400px;
