+++ src/pages/ProductList/Section/Category/FilterCategory/FilterCategory.js
@@ -5,12 +5,9 @@ const FilterCategory = () => {
   return (
     <div className="selectCategory">
       <ul className="selectProductCategory">
-        <li className="productCategoryList">추천순</li>
-        <span className="categoryLine">|</span>
+        <li className="productCategoryListNone">추천순</li>
         <li className="productCategoryList">신상품순</li>
-        <span className="categoryLine">|</span>
         <li className="productCategoryList">낮은 가격순</li>
-        <span className="categoryLine">|</span>
         <li className="productCategoryList">높은 가격순</li>
       </ul>
     </div>
