+++ src/pages/ProductList/Section/Category/FilterCategory/FilterCategory.scss
@@ -9,14 +9,24 @@
     position: relative;
     margin-left: 900px;
     display: flex;
+
     -webkit-box-align: center;
     align-items: center;
 
+    .productCategoryListNone {
+      justify-content: flex-end;
+      margin-left: 8px;
+      padding-left: 10px;
+      white-space: nowrap;
+      color: rgb(140, 140, 140);
+    }
     .productCategoryList {
       justify-content: flex-end;
       margin-left: 8px;
+      padding-left: 10px;
       white-space: nowrap;
       color: rgb(140, 140, 140);
+      border-left: 2px solid rgb(140, 140, 140);
     }
 
     .productCategoryList:hover {
