+++ src/pages/ProductList/Section/Category/ProductCategory/PRODUCTSORTS.js
@@ -1,0 +1,7 @@
+export const PRODUCTSORTS = [
+  { id: 1, text: '전체', name: 'all' },
+  { id: 2, text: '돼지', name: 'pork' },
+  { id: 3, text: '소', name: 'beef' },
+  { id: 4, text: '닭', name: 'chicken' },
+  { id: 5, text: '오리', name: 'duck' },
+];
