+++ src/pages/ProductList/Section/Category/ProductCategory/ProductCategory.js
@@ -1,22 +1,16 @@
 import React from 'react';
 import FilterCategory from '../FilterCategory/FilterCategory';
+import { PRODUCTSORTS } from './PRODUCTSORTS';
 import './ProductCategory.scss';
 
 const ProductCategory = ({ categoryId, changeCategory, products }) => {
-  const PRODUCTS_SORTS = [
-    { id: 1, text: '전체', name: 'all' },
-    { id: 2, text: '돼지', name: 'pork' },
-    { id: 3, text: '소', name: 'beef' },
-    { id: 4, text: '닭', name: 'chicken' },
-    { id: 5, text: '오리', name: 'duck' },
-  ];
-
   return (
     <>
       <ul className="productListTab">
-        {PRODUCTS_SORTS.map(sort => {
+        {PRODUCTSORTS.map(sort => {
           return (
             <li
+              key={sort.id}
               className={`productCategory ${
                 sort.id === categoryId && 'productCategoryActive'
               }`}
