+++ src/pages/ProductList/Section/ProductMeatList/ProductListCard/ProductListCard.scss
@@ -1,4 +1,4 @@
-.productListItem {
+.productListCard {
   margin-top: 50px;
   background-color: rgb(244, 244, 244);
   width: 400px;
